refactor: rename basics to Basic
